requirements-dev.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
DateTime==4.3
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
dj-static==0.0.6
dj.choices==0.11.0
Django==3.2.9
django-allauth==0.44.0
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
urllib3==1.26.7
zope.interface==5.4.0
